Refactor pages by separating logic into hooks
